requirements.txt
fastapi
uvicorn
pillow
numpy
opencv-python
pydantic